+++ src/main/kotlin/model/algorithm/TreeNode.kt
@@ -4,7 +4,7 @@ interface TreeNode<T> {
 
     suspend fun getChildrenWithDeltas(): List<Pair<TreeNode<T>, T>>
 
-    fun getScore(): Int
+    fun getScore(): Double
 
     val isTerminal : Boolean
